split out more pieces of background
